requirements.txt
diskcache==5.6.3
Jinja2==3.1.4
llama_cpp_python==0.3.5
MarkupSafe==3.0.2
numpy==2.2.0
typing_extensions==4.12.2
